fix delete icon listener glitch
make sure the delete icon listener (which is applied when the user
clicks to expand the view) is applied to the correct item; the item's
id is used to save it as the parent element of the delete icon, then the
delete icon is grabbed using the lastChild selector.

previously the icon was selected using the '.delete' class, resulting in
glitches if there were multiple items expanded at once.